requirements.txt
async-generator==1.10
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
exceptiongroup==1.1.0
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
gspread==5.7.2
gspread-dataframe==3.3.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
pandas==1.5.3
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.14
wsproto==1.2.0
